refactor to align with Debian11 edits

+++ Ubuntu_20.04/moab-install.sh
@@ -3,17 +3,15 @@
 ################################################################################
 #!/bin/bash
 set -ex
-if [ "x" == "$1x" ]; then
-	ccores=1
-else
-	ccores=$1
-fi
 
 WD=`pwd`
 name=`basename $0`
 package_name='MOAB'
 
-sudo apt-get install --yes gcc\
+#if there is a .done-file then skip this step
+if [ ! -e ${name}.done ]; then
+
+  sudo apt-get install --yes gcc\
 	cmake\
 	make\
 	build-essential\
@@ -29,15 +27,15 @@ sudo apt-get install --yes gcc\
         python3-setuptools\
 	cython3
 
-#if there is a .done-file then skip this step
-if [ ! -e $0.done ]; then
+  #Should we run make in parallel? Default is to use all available cores
+  ccores=`cat /proc/cpuinfo |grep CPU|wc -l`
+  if [ "x$1" != "x" ]; then
+	ccores=$1
+  fi
 
-  cd $HOME
-  mkdir -p openmc
-  cd openmc
-  mkdir -p MOAB
-  cd MOAB
-  git clone  --single-branch --branch 5.3.0 --depth 1 https://bitbucket.org/fathomteam/moab.git
+  mkdir -p $HOME/openmc/MOAB
+  cd $HOME/openmc/MOAB
+  git clone --single-branch --branch 5.3.1 --depth 1 https://bitbucket.org/fathomteam/moab.git
   mkdir -p build
   cd build
   cmake ../moab -DENABLE_HDF5=ON \
@@ -45,19 +43,17 @@ if [ ! -e $0.done ]; then
               -DENABLE_FORTRAN=OFF \
               -DENABLE_BLASLAPACK=OFF \
               -DBUILD_SHARED_LIBS=ON \
-              -DENABLE_PYMOAB=ON \
+              -DENABLE_PYMOAB=OFF \
               -DCMAKE_INSTALL_PREFIX=$HOME/openmc/MOAB
-  make  -j $ccores
-  make install 
-  cd pymoab
-  bash install.sh
-  sudo python3 setup.py install
+  make -j $ccores
+  make install
+
+  #cd pymoab
+  #bash install.sh
+  #sudo python3 setup.py install
 
   cd ${WD}
-
-  touch ${0}.done
-
+  touch ${name}.done
 else
-   name=`basename $0`
-   echo MOAB appears to already be installed \(lock file ${name}.done exists\) - skipping.
+  echo MOAB appears to be already installed \(lock file ${name}.done exists\) - skipping.
 fi
